pyproject.toml
[project]
name = "discprojagildevufg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.7.9",
    "google-generativeai>=0.8.5",
    "playwright>=1.55.0",
]
